requirements.txt
pandas
PyMuPDF
google-genai
pydantic
python-dotenv